Refactor [#72] 리뷰 반영하여 serviceDto로 리팩토링
